Moved Validators to their respective plugins
