--- plugins/de/opendiabetes/vault/plugin/container/package-info.java
@@ -1,5 +1,0 @@
-/**
- * This package contains enumerations and data structures solely used by plugins.
- * Namely there are only objects used by the {@link de.opendiabetes.vault.plugin.importer.medtronic.MedtronicImporter}.
- */
-package de.opendiabetes.vault.plugin.container;
