+++ plugins/de/opendiabetes/vault/plugin/importer/googlefit/package-info.java
@@ -1,4 +1,5 @@
 /**
- * This package holds this GoogleFitCSVImporter plugin.
+ * This package holds the GoogleFitCSVImporter plugin
+ * and its specifically needed classes.
  */
 package de.opendiabetes.vault.plugin.importer.googlefit;
