+++ plugins/de/opendiabetes/vault/plugin/importer/libretext/package-info.java
@@ -1,4 +1,5 @@
 /**
- * This package contains the LibreText importer plugin.
+ * This package contains the LibreText importer plugin
+ * and its specifically needed classes.
  */
 package de.opendiabetes.vault.plugin.importer.libretext;
