+++ plugins/de/opendiabetes/vault/plugin/importer/medtronic/MedtronicAlertCodes.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package de.opendiabetes.vault.plugin.container;
+package de.opendiabetes.vault.plugin.importer.medtronic;
 
 /**
  * This enum defines Medtronic alerts.
