+++ plugins/de/opendiabetes/vault/plugin/importer/medtronic/MedtronicAnnotatedVaultEntry.java
@@ -14,10 +14,9 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package de.opendiabetes.vault.plugin.container;
+package de.opendiabetes.vault.plugin.importer.medtronic;
 
 import de.opendiabetes.vault.container.VaultEntry;
-import de.opendiabetes.vault.plugin.importer.validator.MedtronicCSVValidator;
 
 import java.util.Objects;
 
@@ -89,7 +88,7 @@ public class MedtronicAnnotatedVaultEntry extends VaultEntry {
         if (!super.equals(obj)) {
             return false;
         }
-        return this.rawType.equals(((de.opendiabetes.vault.plugin.container.MedtronicAnnotatedVaultEntry) obj).getRawType());
+        return this.rawType.equals(((MedtronicAnnotatedVaultEntry) obj).getRawType());
     }
 
 }
