+++ plugins/de/opendiabetes/vault/plugin/importer/medtronic/MedtronicImporter.java
@@ -20,10 +20,7 @@ import com.csvreader.CsvReader;
 import de.opendiabetes.vault.container.VaultEntry;
 import de.opendiabetes.vault.container.VaultEntryAnnotation;
 import de.opendiabetes.vault.container.VaultEntryType;
-import de.opendiabetes.vault.plugin.container.MedtronicAlertCodes;
-import de.opendiabetes.vault.plugin.container.MedtronicAnnotatedVaultEntry;
 import de.opendiabetes.vault.plugin.importer.CSVImporter;
-import de.opendiabetes.vault.plugin.importer.validator.MedtronicCSVValidator;
 import org.pf4j.Extension;
 import org.pf4j.Plugin;
 import org.pf4j.PluginWrapper;
