+++ plugins/de/opendiabetes/vault/plugin/importer/medtronic/package-info.java
@@ -1,4 +1,5 @@
 /**
- * This package holds this MedtronicImporter plugin.
+ * This package holds the MedtronicImporter plugin
+ * and its specifically needed classes.
  */
 package de.opendiabetes.vault.plugin.importer.medtronic;
