+++ plugins/de/opendiabetes/vault/plugin/importer/mysugr/MySugrCSVValidator.java
@@ -14,10 +14,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package de.opendiabetes.vault.plugin.importer.validator;
+package de.opendiabetes.vault.plugin.importer.mysugr;
 
 import com.csvreader.CsvReader;
 import de.opendiabetes.vault.container.VaultEntryAnnotation;
+import de.opendiabetes.vault.plugin.importer.validator.CSVValidator;
 import de.opendiabetes.vault.plugin.util.TimestampUtils;
 
 import java.io.IOException;
