+++ plugins/de/opendiabetes/vault/plugin/importer/mysugr/package-info.java
@@ -1,4 +1,5 @@
 /**
- * This package holds this MySugrCsvImporter plugin.
+ * This package holds the MySugrCsvImporter plugin
+ * and its specifically needed classes..
  */
 package de.opendiabetes.vault.plugin.importer.mysugr;
