+++ plugins/de/opendiabetes/vault/plugin/importer/sony/package-info.java
@@ -1,4 +1,5 @@
 /**
- * This class contains the SonySWR21 importer plugin.
+ * This class contains the SonySWR21 importer plugin
+ * and its specifically needed classes.
  */
 package de.opendiabetes.vault.plugin.importer.sony;
