+++ plugins/de/opendiabetes/vault/plugin/importer/vaultodv/package-info.java
@@ -1,4 +1,5 @@
 /**
- * This package holds this VaultODVImporter plugin.
+ * This package holds the VaultODVImporter plugin
+ * and its specifically needed classes.
  */
 package de.opendiabetes.vault.plugin.importer.vaultodv;
